feat : search components path move
