+++ src/pages/search/RecentSearches.jsx
@@ -1,9 +1,5 @@
 import React from 'react';
-import {
-  BookListSection,
-  ChipsContainer,
-  SectionTitle,
-} from '../../pages/search/Search.styles';
+import { BookListSection, ChipsContainer, SectionTitle } from './Search.styles';
 import { BookChip } from './BookChip';
 
 export const RecentSearches = ({ recentSearches, onClearAll, onRemove }) => {
